Promote phrase metadata helpers to reasoning
