+++ docs/extraction/coordination/queue.md
@@ -1,6 +1,6 @@
 # Upcoming Queue
 
-Last updated: 2026-05-17T15:35Z by codex-2026-05-17
+Last updated: 2026-05-17T16:02Z by codex-2026-05-17
 
 Sequence reflects dependencies. Claim a slice (set Owner) before starting code so a parallel session does not pick the same one. See [`../COORDINATION.md`](../COORDINATION.md) for protocol details.
 
@@ -10,4 +10,5 @@ A-series (cost-closure, `extracted_llm_infrastructure`) is fully merged: PR-A1 #
 | Slice | Product | Owner | Dependencies | Notes |
 |---|---|---|---|---|
 | PR-C2-current-state-reset | `extracted_reasoning_core` | codex-2026-05-16 | Latest merged reasoning-core work: #163 | Reconcile current main against the 2026-05-03 reasoning boundary audit. Current main already has public APIs, semantic cache keys/ports, pack registry, graph/state ports, content-pipeline wrappers, atlas wrappers, and import guards. Next code slice should target a remaining gap, not repeat PR-C1/PR-C4 work. |
-| PR-C3-enrichment-pack-split | `extracted_reasoning_core` | codex-2026-05-17 | PR-C2-current-state-reset | Recommended next code slice: carve atlas-side per-review enrichment into an explicit product pack. |
+| PR-C3-enrichment-pack-split | `extracted_reasoning_core` | merged #564 | PR-C2-current-state-reset | Atlas-side per-review enrichment now lives in an explicit product pack. |
+| PR-C4-phrase-metadata-utility | `extracted_reasoning_core` | codex-2026-05-17 | PR-C3-enrichment-pack-split | Promote pure phrase metadata readers into `atlas_brain.reasoning.phrase_metadata` and leave the task module as a compatibility wrapper. |
