+++ tests/test_atlas_reasoning_evidence_engine_aliases.py
@@ -8,7 +8,7 @@ mixes in the per-review enrichment pack methods (``compute_urgency`` /
 ``override_pain`` / ``derive_recommend`` / ``derive_price_complaint``
 / ``derive_budget_authority``) that stay atlas-side because
 ``derive_price_complaint`` depends on atlas-only
-``_b2b_phrase_metadata``.
+``atlas_brain.reasoning.phrase_metadata``.
 
 These tests pin:
   - atlas's ``EvidenceEngine`` is a subclass of core's slim engine
